services/core-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.26.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
supabase==2.3.0
pytest==8.0.0
pytest-asyncio==0.23.2
pytest-mock==3.12.0